user image

+++ mysite/articles/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views.decorators.http import require_POST
 from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 
 from .models import Article,Comment
 from .forms import ArticleForm,CommentForm
@@ -26,10 +27,15 @@ def detail(request, article_pk):
 @login_required
 def create(request):
     if request.method == "POST":
-        form = ArticleForm(request.POST)
+        form = ArticleForm(request.POST, request.FILES) # form : forms.ModelForm > ArticleForm 클래스의 instance
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False) # article : models.Model > Article 클래스의 instance
+            article.user = request.user
+            article.save()
+            messages.success(request,'게시글 작성 완료')
             return redirect('articles:detail', article.pk)
+        else:
+            messages.error(request,'잘못된 데이터를 넣었따')
     else:
         form = ArticleForm()
     context = {
@@ -55,24 +61,26 @@ def update(request, article_pk):
 def delete(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     if request.method == "POST":
-        article.delete()
-        return redirect('articles:index')
+        if request.user == article.user:
+            article.delete()
+            return redirect('articles:index')
+        else:
+            return redirect('articles:detail', article.pk)
     return redirect('articles:detail', article.pk)
 
 @login_required
 def comment(request,article_pk):
-    # article = get_object_or_404(Article, pk=article_pk)
+    article = get_object_or_404(Article, pk=article_pk)
     # comment = get_object_or_404(Comment, pk=article_pk)
     # comment = Comment.objects.get(article_id=article_pk)
-    article = Article.objects.get(pk = article_pk)
+    # article = Article.objects.get(pk=article_pk)
     if request.method == "POST":
         form = CommentForm(request.POST)
         if form.is_valid():
-           
             comment = form.save(commit=False)
             comment.article = article 
+            comment.user = request.user
             comment.save()
-           
             # comment.content = request.POST.get('content')
             # comment.save()
             return redirect('articles:detail', article.pk)
@@ -84,8 +92,12 @@ def comment(request,article_pk):
     return render(request, 'articles/form.html', context)
 
 @require_POST
-def comment_delete(request,article_pk,comment_pk):
+def comment_delete(request, article_pk,comment_pk):
     comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
-    return redirect('articles:detail',article_pk)
+    if request.user == comment.user:
+        comment.delete()
+    else:
+        pass
+
+    return redirect('articles:detail', article_pk)
